changed int to double

+++ includes/Genetic_Algorithm.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int Evaluate_Circuit(int *circuit_vector, double tolerance, int max_iterations, int num_units, double conc, double waste, double profit, double cost);
+double Evaluate_Circuit(int *circuit_vector, double tolerance, int max_iterations, int num_units, double conc, double waste, double profit, double cost);
